implement a delay for the respawn
after the player dies, they are respawned into the scene by reloading the scene. this commit has delayed the function taht reloads the scene by a defined defined value "restartDelay". you can change this value by selecting the Game Manager, and changing the number in the inspector that corresponds to the "restartDelay" variable.