implement a game "win"
when the player pushes the sphere into the lava, the the game manager will set the bool variable "enemyIsDead" to true when the sphere collides with the lava. then the game conclusion text is decided on based on the enemyIsDead bool ("you win"). then the reset function is called to reload the scene