implement split screen two player
there are now two rigged and animated players in the game.
when one player touches the lava, the screen displays the winner, and the scene resets
I changed the type of movement that the players use. its abhorrent, but it was necessary in order to make the game playable on one keyboard. next step is to map everything to xbox controllers because we have some handy